Separate SSZType part
